requirements.txt
Django==1.5.5
Jinja2==2.7.2
MarkupSafe==0.21
Pillow==2.0.0
PyYAML==3.11
Pygments==1.6
South==0.8.4
Sphinx==1.2.2
Unidecode==0.04.16
argparse==1.2.1
bleach==1.4
crispy-forms-foundation==0.3.4
diff-match-patch==20110725.1
dj-database-url==0.2.2
dj-static==0.0.5
django-absolute==0.3
django-ajax-selects==1.3.4
django-appconf==0.6
django-autoslug==1.7.2
django-avatar==2.0
django-bower==5.0.1
django-ckeditor-updated==4.2.6
django-crispy-forms==1.4.0
django-debug-toolbar==1.0
git+https://github.com/toastdriven/django-haystack@cddc4db6e424dc66f1963aa8948ea864
django-extensions==1.3.8
django-haystack==2.1.1-dev
django-leaflet-storage==0.5.0
django-registration==1.0
django-toolbelt==0.0.1
docutils==0.11
graphviz==0.3.4
gunicorn==18.0
html5lib==0.999
mock==1.0.1
psycopg2==2.5.1
pygraphviz==1.2
shortuuid==0.4.2
simplejson==3.5.2
six==1.5.2
sqlparse==0.1.10
static==0.4
wsgiref==0.1.2
